brain_tumor_detection/Backend/requirements.txt
Flask
Flask-CORS
tensorflow
keras
numpy
pandas
opencv-python
scikit-learn
matplotlib
Pillow
gunicorn
